2009-09-08  Kai Tietz  <kai.tietz@onevision.com>
        * tree-ssa-reassoc.c (find_operand_rank): Cast pointer
        via intptr_t to long type.
        (insert_operand_rank): Cast long type via intptr_t to
        pointer type.
        * genattrtab.c (RTL_HASH): Use intptr_t to cast from
        pointer to scalar.
        * c-pretty-print.c (pp_c_tree_decl_identifier): Cast
        from pointer to unsigned via uintptr_t.



git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@151515 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/c-pretty-print.c
@@ -2235,7 +2235,7 @@ pp_c_tree_decl_identifier (c_pretty_printer *pp, tree t)
   else
     {
       static char xname[8];
-      sprintf (xname, "<U%4x>", ((unsigned)((unsigned long)(t) & 0xffff)));
+      sprintf (xname, "<U%4x>", ((unsigned)((uintptr_t)(t) & 0xffff)));
       name = xname;
     }
